bug 911 / fixed
HTML changes:
src/job-owners/index.html

adding caption tag (with a description ot the table) after every open
table tag, and removing the caption tag that wrap's the
"h2-wrapper".

In every TD tag that doesn’t contain A tag in it,
Wrap
the content in a span tag and give it the attributes tabindex=“0” and
role=“presentation”.

Adding tabindex=“0” and role=“presentation”
attributes to the relevant elements in the "mobil-toggle-wrapper" as
well. .